pyproject.toml
[project]
name = "endoreg-db"
version = "0.5.3"
description = "EndoReg Db Django App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agl-report-reader>=0.4.0",
    "django-bootstrap5>=24.3",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "faker>=32.1.0",
    "ffmpeg-python>=0.2.0",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "pytorch-lightning>=2.4.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "tesseract>=0.1.3",
    "tqdm>=4.67.0",
    "urllib3>=2.2.3",
]

[tool.setuptools]
packages = ["endoreg_db"]
